requirements/base.txt
Markdown>=2.4.1
django-model-utils>=2.2
bcrypt>=1.0.2
psycopg2>=2.5.3
Pillow>=2.8.1
celery>=3.1.17
django-celery>=3.1.16
amqp>=1.4.6
djangorestframework>=3.1.2
